fix(pwa): Corrigida estrura de pastas e imports
