+++ pwa/src/components/DeviceCard/styles.js
@@ -1,1 +1,8 @@
-export default {};
+export default {
+  container: {
+    width: 335
+  },
+  header: {
+    cursor: 'pointer'
+  }
+};
